+++ xActionCablePro/lib/src/base/connection_interceptor.dart
@@ -1,5 +1,7 @@
+import 'dart:async';
+
 import 'package:x_action_cable_pro/src/base/connection_reference.dart';
 
 abstract class ConnectionInterceptor {
-  ConnectionReference beforeConnection(ConnectionReference reference);
+  FutureOr<ConnectionReference> beforeConnection(ConnectionReference reference);
 }