+++ xActionCable/CHANGELOG.md
@@ -1,3 +1,6 @@
+# 2024-07-26 1.0.2
+- ActionCable disconnect posibility
+
 # 2024-06-16 1.0.1
 Initial release