+++ xActionCablePro/CHANGELOG.md
@@ -1,3 +1,12 @@
+# 2024-07-26 1.1.0
+Fix param names and better handling subscriptions
+
+# 2024-07-23 1.0.2
+Support for async interceptors
+
+# 2024-07-22 1.0.1
+Set correct versions
+
 # 2024-06-16 1.0.0
 Initial release