+++ xActionCableProGen/CHANGELOG.md
@@ -1,3 +1,12 @@
+# 2024-07-26 1.0.3
+Update dependency x_action_cable_pro to ^1.1.0 and rename channelParams to params
+
+# 2024-07-23 1.0.3
+Update dependency x_action_cable_pro to ^1.0.2
+
+# 2024-07-22 1.0.2
+Update dependencies
+
 # 2024-07-17 1.0.1
 Release fix